// fe/services/types.ts
export type gaugeFormValues = {
    gaugeSt: number;
    gaugeWidth: number;
    finalWidth: number;
};

export type calculatorFormValues = {
    currentSt: number;
    numberSt: number;
};

export type Yarn = {
    name: string;
    grams: number;
    yardage: number;
    companyName: string;
    weight: string;
    gaugeSt: number;
    gaugeIn: number;
    fibers: string;
    texture: string;
}